Alterações e adições lógicas para o jogo
Adição de mais classes para manipular os dados, alteração na forma de calculo da pontuação final e reajustes

+++ SuperHeroBuilder/Builders/SuperHeroBuilder.cs
@@ -53,13 +53,13 @@ namespace SuperHeroBuilder.Builders
             return this;
         }
 
-        public ISuperHeroBuilder BuildAttack(double attack)
+        public ISuperHeroBuilder BuildAttack(int attack)
         {
             _superHero.Attack = attack;
             return this;
         }
 
-        public ISuperHeroBuilder BuildDefense(double defense)
+        public ISuperHeroBuilder BuildDefense(int defense)
         {
             _superHero.Defense = defense;   
             return this;
